requirements.txt
flask
flask-cors
langchain
langchain-google-genai
google-generativeai
chromadb
pypdf
python-dotenv
